refactor(powershell): move step logic out of Powershell
